Move Hex and ZipUtils to util package.
