+++ api/src/main/java/ai/djl/util/Hex.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.repository;
+package ai.djl.util;
 
 /** {@code Hex} is a set of utilities for working with Hexadecimal Strings. */
 public final class Hex {
